Added calling tweens,out of the box roblox support
Calling tweens are now scripts that are sent values between 0 and 1, you can use a lerp function to act like a tween. It's great for modifying userdata with a tween.

I added out of the box Roblox support at line 404, it checks if it's running in Roblox. If so, it starts a thread that automatically calls tween.update.